more refactoring int->int64

+++ examples/wba/_01_assembling_wbs.py
@@ -49,7 +49,7 @@ def window(start, end, **parameter):
 
 def naive_stride_list(start, stop, duration, foot=None, **paras):
     """A window list full of identical strides."""
-    x = np.arange(start, stop + duration, duration)
+    x = np.arange(start, stop + duration, duration, dtype="int64")
     start_end = zip(x[:-1], x[1:])
 
     return pd.DataFrame.from_records(
